requirements.txt
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
asgiref==3.7.2
psycopg2==2.9.9
pytz==2023.3.post1
typing_extensions==4.8.0
sqlparse==0.4.4